requirements.txt
langchain
langgraph
langchain-openai
pandas
python-dotenv
ipython
setuptools 